+++ src/main/java/ar/edu/unlam/tallerweb1/servicios/ServicioCalcular.java
@@ -4,5 +4,5 @@ import org.springframework.stereotype.Service;
 
 @Service("servicioCalcular")
 public interface ServicioCalcular {
-	Integer calcular(Integer numero1, String operador, Integer numero2);
+	Double calcular(Double numero1, String operador, Double numero2);
 }
